Added animatedobject for player

+++ BaseGame99/GameObject.h
@@ -5,10 +5,10 @@ public:
 	GameObject();
 	~GameObject();
 
-	float x_location;
-	float y_location;
-	float x_speed;
-	float y_speed;
+	int x_location;
+	int y_location;
+	int x_speed;
+	int y_speed;
 
 	ALLEGRO_BITMAP* bmp;
